backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
opencv-python
mediapipe
requests
numpy
replicate
python-dotenv
